refactor: edit dataset hooks (#7061)
