+++ frontend/src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/components/MoreDropdown/components/Menu/types.ts
@@ -1,7 +1,7 @@
 import { Props as ChooserProps } from "src/components/DropboxChooser";
 import { Collection, Dataset } from "src/common/entities";
 import { UseMoreMenu } from "src/views/Collection/hooks/useMoreMenu";
-import { EditDataset } from "src/views/Collection/hooks/useEditCollectionDataset/common/entities";
+import { EditDataset } from "src/views/Collection/hooks/useEditCollectionDataset/types";
 
 export interface MenuItemProps {
   editDataset: EditDataset;
