+++ frontend/src/views/Collection/hooks/useCollectionDatasetForm/utils.ts
@@ -1,0 +1,82 @@
+import {
+  FieldErrors,
+  FieldValues,
+} from "src/views/Collection/hooks/useCollectionDatasetForm/types";
+
+/**
+ * Returns field values and errors from form data.
+ * @param formEl - Form element.
+ * @param defaultValues - Default field values.
+ * @returns field values and errors tuple.
+ */
+export function getAndValidateFieldValues(
+  formEl: HTMLFormElement,
+  defaultValues: FieldValues
+): [FieldValues, FieldErrors] {
+  const fieldValues = getFieldValues(formEl, defaultValues);
+  const fieldErrors = validateForm(fieldValues, defaultValues);
+  return [fieldValues, fieldErrors];
+}
+
+/**
+ * Returns field values from form data.
+ * @param formEl - Form element.
+ * @param defaultValues - Default field values.
+ * @returns field values.
+ */
+export function getFieldValues(
+  formEl: HTMLFormElement,
+  defaultValues: FieldValues
+): FieldValues {
+  const formData: FormData = new FormData(formEl);
+  const fieldValues = {} as FieldValues;
+  for (const key of Object.keys(defaultValues)) {
+    if (formData.has(key)) {
+      const value = formData.get(key);
+      Object.assign(fieldValues, { [key]: formatFormDataValue(value) });
+    }
+  }
+  return fieldValues;
+}
+
+/**
+ * Returns form data entry value, formatted, if necessary.
+ * @param value - Form data entry value.
+ * @returns formatted data entry value.
+ */
+export function formatFormDataValue(
+  value: FormDataEntryValue | null
+): FormDataEntryValue | null {
+  if (typeof value === "string") {
+    return value.trim();
+  }
+  return value;
+}
+
+/**
+ * Returns true if the form data is valid.
+ * @param errors - Field errors.
+ * @returns true if the form data is valid.
+ */
+export function isValid(errors: FieldErrors): boolean {
+  return Object.keys(errors).length === 0;
+}
+
+/**
+ * Validates form data.
+ * @param fieldValues - Field values.
+ * @param defaultValues - Default field values.
+ * @returns error.
+ */
+export function validateForm(
+  fieldValues: FieldValues,
+  defaultValues: FieldValues
+): FieldErrors {
+  const fieldErrors = {} as FieldErrors;
+  for (const [key, value] of Object.entries(fieldValues)) {
+    if (!value || value === defaultValues[key]) {
+      Object.assign(fieldErrors, { [key]: `${key} is required` });
+    }
+  }
+  return fieldErrors;
+}
