+++ frontend/src/views/Collection/hooks/useEditCollectionDataset/useEditCollectionDataset.ts
@@ -1,45 +1,50 @@
-import {
-  FieldValues,
-  UseEditCollectionDataset,
-} from "src/views/Collection/hooks/useEditCollectionDataset/common/entities";
+import { UseEditCollectionDataset } from "src/views/Collection/hooks/useEditCollectionDataset/types";
 import { useEditDataset } from "src/common/queries/collections";
 import {
+  FieldValues,
   PathParams,
   SubmitOptions,
-} from "src/components/Collection/components/CollectionDatasetsGrid/components/Row/DatsetRow/components/EditDataset/hooks/types";
+} from "src/views/Collection/hooks/useCollectionDatasetForm/types";
+import { useCallback } from "react";
+import { useCollectionDatasetForm } from "src/views/Collection/hooks/useCollectionDatasetForm/useCollectionDatasetForm";
 
 /**
  * Edit functionality for collection dataset.
  */
 export function useEditCollectionDataset(): UseEditCollectionDataset {
+  const editDatasetFormMethod = useCollectionDatasetForm();
   const editDatasetMutation = useEditDataset();
 
-  const onEditDataset = async (
-    pathParams: PathParams,
-    fieldValues: FieldValues,
-    submitOptions?: SubmitOptions
-  ): Promise<void> => {
-    // Send order to BE.
-    const { collectionId, datasetId } = pathParams;
-    const payload = JSON.stringify(fieldValues);
-    await editDatasetMutation.mutateAsync(
-      {
-        collectionId,
-        datasetId,
-        payload,
-      },
-      {
-        onSuccess: () => {
-          submitOptions?.onSuccess?.();
+  const onEditDataset = useCallback(
+    async (
+      pathParams: PathParams,
+      fieldValues: FieldValues,
+      submitOptions?: SubmitOptions
+    ): Promise<void> => {
+      // Send order to BE.
+      const { collectionId, datasetId } = pathParams;
+      const payload = JSON.stringify(fieldValues);
+      await editDatasetMutation.mutateAsync(
+        {
+          collectionId,
+          datasetId,
+          payload,
         },
-        onError: () => {
-          submitOptions?.onError?.();
-        },
-      }
-    );
-  };
+        {
+          onSuccess: () => {
+            submitOptions?.onSuccess?.();
+          },
+          onError: () => {
+            submitOptions?.onError?.();
+          },
+        }
+      );
+    },
+    [editDatasetMutation]
+  );
 
   return {
     editDatasetAction: { onEditDataset },
+    formMethod: editDatasetFormMethod,
   };
 }
